refactor: PathManager를 인터페이스로 변경 및 원래 구현체를 DijkstraPathManager로 분리
